requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
